Wait for signature and attestation
In the chains test we before we check the attestation and signature we
need to make sure that the chains-controller has had enough time to
create them. Otherwise we would fail the cosign* tasks due to the images
with the attestation and signature not pushed/tagged in the interlan
registry.

This adds a step to wait for a maximum of 60 seconds for the images to
be created by listing the ImageStreamTag and looking for the specific
`.att` and `.sig` image tags to be present there.

Fixes https://issues.redhat.com/browse/HACBS-278