feat: use pipeline image from config map
The `build-pipelines-defaults` ConfigMap points to the current pipeline
bundle, so instead of a fixed bundle image we should use the value
placed there. This also allows running the end to end tests against a
set of bundle images from a personal quay.io registry.

We're now starting PipelineRuns instead of TaskRuns and we need the
change in redhat-appstudio/build-definitions#112 merged for this to
run successfully.

Ref. https://issues.redhat.com/browse/HACBS-424