fix: Less brittle contract task result assertions
The aim here is to avoid breaking the e2e tests when we vary the
conftest output format, e.g. to add additional keys or revise the
wording of the failure messages.

Uses the new tekton.MatchTaskRunResultWithTemplate matcher defined
in the previous commit.

Note that there can only be one match per task result in
ContainElements, so that's why we need to do all the assertions in
one single matcher.