fix: everything in Enterprise contract tests suite (#151)
* refactor: tests for `cosign-verify-*` tasks

Follow up on redhat-appstudio/build-definitions#173 to remove testing of
`cosign-verify-*` tasks that are removed there.

* refactor: bundle image reference for pull requests

Pull request bundle images are from
`quay.io/repository/redhat-appstudio/pull-request-builds` repository and
they have tags reflecting their flavor (base or HACBS). The replacement
that was looking for `/<flavor>` was not appropriate for this as the
flavor was in the tag (e.g. `:<flavor>`). This adjusts for both pull
request builds that change the bundles and those that do not.
It also seems that `HACBSCoreServiceTemplatesBundle` is unused, so
instead of adjusting that the logic to deduce the value and the field
has been removed.

* feat: labels for test suites

This sets the default labels on suites and all tests within those. This
allows us to selectively run or disable whole test suites.

For example to disable running Enterprise contract test suite, labeled
with `ec`, run with:

    --ginkgo.label-filter='!ec'

* refactor: remove skip functionality in ec tests

With the default labels being set at the Enterprise contract test suite
level, we no longer need the skip functionality within the test suite,
as we can skip the test by excluding the label in the `label-filter`
parameter, i.e. run with `--ginkgo.label-filter='!ec'`.

* refactor: use . import for ginkgo

For nicer readability of tests.

* refactor: cleanup debug messages and assertions

Debug are now more descriptive and in correct places. The assertions
have their `expected` and `received` values in correct places.

* feat: use EnterpriseContractPolicy custom resource

This refactors the tests to create and use EnterpriseContractPolicy
custom resource instead of the ConfigMap. In verify-enterprise-contract
Tekton Task both ConfigMap and EnterpriseContractPolicy are supported,
whereas in v2 only EnterpriseContractPolicy is supported.

The tests now assert the state as expected today, i.e. the failing Tasks
are asserted to fail as we expect today. We no longer assert the
`OUTPUT` TaskRun result as that is brittle with corresponding changes to
the changes in hacbs-contract/ec-policy repository.

By default the policy sources are fetched from hacbs-contract/ec-policy
repository main branch. This can be overridden by creating a ConfigMap
e2e-tests/ec-config with keys `revision` and `repository`, each of those
can replace this default settings. This will come in useful if we want
to pin a certain revision we know the tests are good to run against,
e.g. if we enable pull request updates via
redhat-appstudio/infra-deployments from hacbs-contract/ec-policies; or
if we setup a separate repository with policies suited for the end to
end tests. This brings the groundwork for that, either of those options
could be a good solution to reduce the brittleness, but this is left for
the future.

Also refactors the tests to add more debug output when the tests fail.

* refactor: move initialization call to BeforaAll