feat: JSONPath matching for TaskRunResults
This enables asserting parts of the TaskRunResult values with JSONPath,
assuming that those contain a JSON value.

The JSONPath is in the syntax `{...}` and the asserted value needs to
conform to an array of values. The array of values is needed because the
evaluation of the of the JSONPath might yield more than one value.