fix: two layers needed in attestation image
The Tekton Chains controller will push two layers to the attestation
image. One containing the TaskRun attestation and one containing the
PipelineRun attestation. We require PipelineRun attestation to be
present. Without downloading and examining the image we can make sure
that the PipelineRun attestation is present if the `ImageStreamTag` has
two `dockerImageLayers`.

We're now fetching the `ImageStreamTag` directly rather than listing all
and filtering the items list for `ImageStreamTag` with a specific name.