fix: disable release tests
A backwards incompatible change was done in the CRD
EnterprisePolicyConfig. Changes in the e2e tests and infra-deployments
must happen at the same time in order to proceed. Updating one without
the other causes CI to fail. Since updating both at the same time is not
possible, the release tests are disabled. This was also done for the
EC/Chains tests.

Once the changes are merged in infra-deployments, we can change the
tests to accommodate the new CRD format and re-enable them.

https://issues.redhat.com/browse/HACBS-1525

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>