chore: disable the verbose output (#297)
* chore: disable the verbose output

Gikngo should display the output accumulated in `GinkgoWriter` when a
scenario fails, otherwise the output is quite verbose and displayed for
both failing and succeeding scenarios.

* chore: use colored output

I'm assuming most folk will interact with the e2e tests from the prow
web interface. In prow ansi color is supported and I think the use of
will help folk to contextualize/understand the output easier.

* chore: replace usage of klog with giknkgowriter

To have the debug lines printed only when a scenario fails this opts to
use the `GiknkgoWriter` instead of `klog`. See[1]

[1] https://onsi.github.io/ginkgo/#logging-output

* chore: forward klog to ginkgowriter

This forwards messages passed onto `klog` to the `GinkgoWriter` via
`logr` implementation. The environment variable `KLOG_VERBOSITY` can be
set to a number 0..9 to control the `klog` logging level.

* chore: update readme