refactor: use fixed versions for release pipeline bundles (#491)
As agreed on today's HACBS Release Retrospective call,
let's switch to using fixed versions when referencing
pipeline bundles.

This will avoid situations like the one we hit recently
in RHTAPBUGS-197 when a new task was added to a new
version of a pipeline and this was immediately used
in e2e tests and it failed (due to a bug in the new
task).

Going forward, we'll update the version in e2e tests
as needed. Typically we'll have an epic for a new feature
and updating e2e tests will be one of the stories there.

Signed-off-by: Martin Malina <mmalina@redhat.com>