feat(HACBS-2222): update release pipeline versions
This will update the release tests to use the latest
versions of the release and push-to-external-registry
pipelines. The new versions use the new failOnEmptyResult
parameter when using the apply-mapping task.

As a result, when the mapping results in an empty component
list, the apply-mapping task will fail in the pipeline.
This should make debugging of failed e2e tests a little
more convenient - this is something that happened repeatedly
in the last month and it's often not easy to realize
what's going on - until now, the typical symptom would be
a failed EC test - but the rest of the pipeline would happily
pass.

Signed-off-by: Martin Malina <mmalina@redhat.com>