feat: HACBS-2315: Remove outdated EC dependency (#618)
The ec-defaults ConfigMap in the enterprise-contract-service namespace
contain values for which policy and data refs to use when creating a new
EnterpriseContractPolicy resource. However, we have moved away from
using this ConfigMap for this purpose. Although the values still exist,
they are no longer updated by automation.

The automation now keeps a few EnterpriseContractPolicy resources up to
date. This commit changes the release tests to no longer fetch the
policy and data refs from the ConfigMap. Instead, it uses the `default`
EnterpriseContractPolicy as the basis for creating a new resource. This
`default` has the added benefit of only including certain policy rules
that shouls always pass under normal operation.

NOTE: The ConfigMap is still useful for determining the most up to date
ref for the ec Tekton Task. There are no plans to change this.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>