fix: Do not require attestations to have 2 layers
With HACBS-2209, Tekton Chains will only produce a single attestation,
for the PipelineRun. This is expected. This change updates the tests to
no longer assert that the attestation image has 2 layers. This is an
implementation detail anyways. It is better asserted indirectly by
testing the actual functionality of Enterprise Contract.

Note that this assertion has failed before due to a different issue
which is explained in the second comment in RHTAPBUGS-375. It is
expected that this patch will also address that issue by making the
originaly reported underlying problem more obvious. There the
PipelineRun failed for a different reason.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>