refactor(HACBS-2375): adjust release test to pyxis changes
Pyxis containerImage object will no longer
contain the embedded contentManifestComponent objects.

If we really still wanted to check that there is at least
one component in Pyxis, we could use graphql and edges
to get this information.

But it's probably ok for us to just check that a manifest
is included in the image - that still confirms that
upload_sbom script did something. If the upload_sbom
script didn't run, the manifest id would be empty.

Signed-off-by: Martin Malina <mmalina@redhat.com>