fix: upgrade ec release policy
This upgrades EC release policy to the latest version which includes a
fix[1] needed for the upgraded version of OPA in[2].

The proper fix would not use a pinned version of rules/data but rely on
the defaults from the `enterprise-contract/default` ConfigMap, as used
in other scenarios. This is not possible at the moment because we need
a change[3] that restores the missing acceptable bundles merged first.

[1] https://github.com/enterprise-contract/ec-policies/pull/546
[2] https://github.com/enterprise-contract/ec-cli/pull/801
[3] https://github.com/redhat-appstudio/infra-deployments/pull/2130