feat: test that EC Task is run in Release pipeline
In [HACBS-2472](https://issues.redhat.com/browse/HACBS-2472) we realized
a gap in the testing putting the Release pipeline at risk if the
`verify-enterprise-task` changes in an incompatible way with the Release
pipeline.

This adds a coverage to test that `verify-enterprise-task` is included
in the Release PipelineRun and that it has succeeded.

The success is assumed by the release PipelineRun succeeding, though
there is no check to make sure that `verify-enterprise-task` is included
and that it succeeds individually.

Currently it is not possible to ignore errors from Tasks in PipelineRuns
so this safeguards in case TEP-0050 gets implemented.

Ref. https://issues.redhat.com/browse/HACBS-2472

[1] https://github.com/tektoncd/community/blob/fe0c2bbedb0754cd6b4352eab5d344683a73c296/teps/0050-ignore-task-failures.md