feat: Test EC pipelines
The EC pipelines from the build-definitions repository are used by the
auto-created IntegrationTestScenario. This adds test coverage to those
Tekton pipelines.

By default it tests the pipelines from the main branch of the
build-definitions repo. The branch as well as the repository url can be
overwritten via environment variables. This will be used in the future
to test pull requests in the build-definitions repository.

https://issues.redhat.com/browse/HACBS-2475

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>