feat: dump logs on failure (#714)
feat: dump relevant logs on failure

This PR will dump logs to the main build log that correspond to the time
that the failed test was running. This should give all logs in once
place, and remove the need to hunt though multiple log files while
manually correlating time stamps.