fix: use custom label for ec pipelines
The EC pipelines tests are expected to be used by the build-definitions
CI. When it runs there, it runs directly on the prod cluster. The tests
in `tests/enterprise-contract/contract.go` don't work with that setup.
Also, they are not really related to changes in build-definitions.

This commit adds a new label, `ec-pipelines`, so the build-definitions
CI can include just the relevant tests.

https://issues.redhat.com/browse/HACBS-2475

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>