fix: move EC pipelines tests to build-templates
The tests for the EC pipelines are meant to, mostly, be executed from
the build-definitions CI. That runs directly against prod. As such, the
user has no access to things like the quay.io secret. The previous
iteration of these tests relied on such access in order to run a
pipeline for a test repo.

This change makes it so an Application is used to build the test repo.
This removes any sort of access issues. Also, this removes the need to
build a specific test image just for these tests. Instead, the same test
image is used for all the tests.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>