fix: remove ec pipelines after verification
The PipelineRun cleanup job may kick in if there are too many
PipelineRuns in the namespace. If this happens during a test run, it may
delete a PipelineRun for a component that hasn't been tested yet. In
that case the test would fail. This change deletes the EC PipelineRuns
as soon as we're done with it. This lowers the chance of the cleanup job
causing random test failures.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>