fix: move EC pipelines tests to build-templates (#743)
* fix: move EC pipelines tests to build-templates

The tests for the EC pipelines are meant to, mostly, be executed from
the build-definitions CI. That runs directly against prod. As such, the
user has no access to things like the quay.io secret. The previous
iteration of these tests relied on such access in order to run a
pipeline for a test repo.

This change makes it so an Application is used to build the test repo.
This removes any sort of access issues. Also, this removes the need to
build a specific test image just for these tests. Instead, the same test
image is used for all the tests.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* fix: remove ec pipelines after verification

The PipelineRun cleanup job may kick in if there are too many
PipelineRuns in the namespace. If this happens during a test run, it may
delete a PipelineRun for a component that hasn't been tested yet. In
that case the test would fail. This change deletes the EC PipelineRuns
as soon as we're done with it. This lowers the chance of the cleanup job
causing random test failures.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

---------

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>