fix: remove the test we already have coverage for (#758)
This test is testing a case we already have coverage in the repository
where the code resides[1]. We don't want to be testing the behaviour of
EC in e2e tests but integration points between components.

This test started failing because it uses the up-to-date acceptable task
bundles data and in those the policy rule no longer emits a warning
because another rule in the same package emits a violation of using
out-of-date (non acceptable) bundle.

See https://github.com/redhat-appstudio/infra-deployments/pull/2357

Ref. https://issues.redhat.com/browse/HACBS-2592

[1] https://github.com/enterprise-contract/ec-policies/blob/2d4adcbc0befbbc3e04c93a599dba4f89e581e82/policy/release/attestation_task_bundle_test.rego#L89-L107

Co-authored-by: Flavius Lacatusu <flacatus@redhat.com>