feat(RHTAPREL-298): change tests to use new Release API
The API for the Release Service has changed, removing
ReleaseStrategy and relying on ReleasePlanAdmissions
to pass most of the information. There are multiple changes
involved in different repositories, so some pipelines had
to be "disabled" moving them to the pipelines directory.
We will take care of those once all the changes are merged.

Signed-off-by: David Moreno García <damoreno@redhat.com>