fix(RHTAPBUGS-883): make sure revision is set
In fixing RHTAPBUGS-880 we found out that the policy did not correctly
handle undefined source revision, which lead to RHTAPBUGS-883.

Before we merge RHTAPBUGS-883 we need to make sure that the revision is
passed via the parameters to EC so that the change for fixing
RHTAPBUGS-883 doesn't break e2e tests.