fix: remove faulty EC test
Thes test "ec cli supports the current policy config fields" looks at
the output of the verify ec task and looks for the presence of a certain
string. This is taken as a way to verify the policy config is being
propagated to the task properly. This is a weak assertion because the
particular string may appear somewhere else. Also, this is checking a
very specific behavior of the EC cli and not an integration test. There
is already sufficient test coverage on the EC cli's own tests.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>