fix: extend the Chains attestation timeout (#1067)
* feat: one common constant for attestation timeout

Note: this unifies the timeout at 10 minutes everywhere. Previously, it
was 5 minutes in some places.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

* fix: extend the Chains attestation timeout

It can take The Tekton Chains operator a long time to attest images.
Recently, in the build-definitions test suite, we observed Chains taking
nearly 30 minutes - even an hour in extreme cases - to do its job.

We don't have much data at the moment, only these 6 cases (time taken
between the push of the binary image and the push of the attestation,
based on timestamps in quay.io):

    56m
    49m
    14m
    13m
    24m
    27m

A timeout of 30 minutes seems to be necessary to handle the common case
(while still not enough for the worst case).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

---------

Signed-off-by: Adam Cmiel <acmiel@redhat.com>