fix: update EC policy inclusion (#1136)
The policy rule `attestation_task_bundle.task_ref_bundles_acceptable`
was recently renamed to
`attestation_task_bundle.task_ref_bundles_trusted`. This commit updates
the test to include either to facilitate the transition. Once all users
are up to date, we can drop inclusion of the old name.

This change is required because the test relies on this particular
policy rule to fail.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>