fix: use correct EC collection (#1194)
In order to use an EC policy collection, the `@` prefix must be used.
Otherwise, EC assumes the value refers to a package.

There were two tests that were using this convention incorrectly.
Previously, this simply caused EC to ignore it. However, an upcoming
change in EC ensures that at least some rules execute. Otherwise, EC
assumes the user has made a typo in their policy config.

Well... that is working quite well because it has caught a bogus test in
this repo.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>