feat(RELEASE-927): add multi arch test with rh-advisories pipeline (#1258)
feat(RELEASE-927): add multi arch test with rh-advisories release pipeline

Signed-off-by: Jing Qi <jinqi@redhat.com>

In this test case, a dedicated repo with PaC configuration to staging cluster is
 used. The repo is tied to a fixed component in an application which need to be
created before the test case is ran.
The test case uses a push event (merging a PR) to trigger the build pipelineRun.
RH-advisories release pipelineRun is used for the multi-arch test.