refactor(KONFLUX-3773): Fork repositories sequentially
GitHub does not allow forking more than 3 at a time anyway. If 3 are
forking already, it returns 403 for more forking requests and as we are
retrying after 2 seconds, we soon used whole rate_limit.

With parralel forking it also often happened that process was stuck in
loop with 403 and only think that helped was cleaning all 10 temporary
repos like "<source_repo>-1".

Anyway, fingers crossed, this will be slower but more stable.