feat: Test different Pipeline variants (#1384)
* feat: Remove check for show-summary presence

In newer build Pipelines, the show-summary Task is no longer included.
It is also no longer a required Task. As such, this commit removes the
test that ensure the show-summary Task was included in the build
Pipeline.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* feat: Test different Pipeline variants

This commit modifies the structure of the component test scenarios to
allow them to be executed by different build Pipelines.

Ref: https://issues.redhat.com/browse/EC-715

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

* fix: disable new multi-arch tests

Blocked on KFLUXINFRA-836

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

---------

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>