feat: enable dockerfile check for hermetic builds (#1404)
* feat: enable dockerfile check for hermetic builds

STONEBLD-2795

Hermetic builds should now push the original dockerfile, not the
modified one. Enable the check.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

* fix: improve dockerfile check error messages

When the dockerfile image doesn't exist, print the expected image

Compare the content of the Dockerfile as a string, not as an array of
bytes. Otherwise the error message is not readable to humans.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

---------

Signed-off-by: Adam Cmiel <acmiel@redhat.com>