fix: update release multiarch data fields
This commit fixes two data fields in the multiarch advisories release
pipeline test. The product_id field is changed from a string to an int,
and the component source is removed from the data section of the
ReleasePlanAdmission. The same component source information already
appears to be present in the create snapshot function.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>