fix: git repo url should have .git suffix (#1480)
This url is used in the RPA and it's used for Tekton
git resolver. All the RPAs in konflux-release-data repo
use the .git suffix and it's also what's used
as an example in Tekton docs.

For context, a user recently hit an issue:
https://github.com/konflux-ci/release-service-catalog/pull/759

We had a bug that caused an issue with the .git suffix. But e2e
didn't expose the problem because the .git suffix was not there.

Signed-off-by: Martin Malina <mmalina@redhat.com>