feat(RELEASE-1401): add fixed issue in release advisory test
We are adding more and more tasks to the advisory pipeline that do
things based on the issues.fixed in the releaseNotes key. To ensure
testing coverage, this commit adds an issue to the advisory test. The
issue used is one of the ones opened for some of the advisory issue
improvements, so that we do not run the risk of interfering with anyone
else's issues.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>