fix(KONFLUX-7733): advisory_url to point to customer portal (#1574)
* fix(KONFLUX-7733): advisory_url to point to customer portal

We are changing the format of the advisory url that is produced
in the release pipeline. Instead of pointing to the gitlab repo,
it will now point to the customer portal.

Signed-off-by: Martin Malina <mmalina@redhat.com>

* chore: change default release-service-catalog branch to development

Having it set to staging is problematic when you need to make
a change in the catalog that requires a change in e2e. We hit this
recently when working on KONFLUX-7733 .

I was able to pair the changes in both repos and merge the catalog
change. Now I would like to merge the e2e change, but I can't
because the e2e run in the e2e PR will use staging brach of the
catalog where the change is not yet present. And at the same time,
all other catalog PRs will now fail on e2e, because the e2e PR
is not merged yet.

Signed-off-by: Martin Malina <mmalina@redhat.com>

---------

Signed-off-by: Martin Malina <mmalina@redhat.com>