fix: OverrideMediaType to properly update BUILDAH_FORMAT (#1706)
* fix(build_template): don't duplicate BUILDAH_FORMAT param (#1702)

if BUILDAH_FORMAT param already existed in PLR definition, it was
just duplicated and created invalid definition.

Check if param exists and then just update it, if param doesn't exist add
a new one.

This was accidentally reverted in
5c9014c243fdfec0c4c6286499e3bb9a291db718

Signed-off-by: Martin Basti <mbasti@redhat.com>

* fix: missing OverrideMediaType in deepcopy function

This attribute was ignored during copy, thus it was never invoked in the
tests.

Signed-off-by: Martin Basti <mbasti@redhat.com>

---------

Signed-off-by: Martin Basti <mbasti@redhat.com>