fix(KONFLUX-10309): Use right organization when getting templates
Normally we are using this template repo:

    https://github.com/rhtap-perf-test/konflux-probe-test-templates

That lives in same organization as all the component repos we work with.

When Zhiming tried to use his fork with one small change:

    https://github.com/zxiong/konflux-probe-test-templates

getting file from there did not seen the change.

Turns out our code was actually getting the file from `rhtap-perf-test`.

This fixes the issue and as far as the GitHub token is sufficient to get
thta file from that different repo in different organization, it will
try to do so.