SBOM util: Retrieve the name of the tool that created a component
Tools such as Hermeto and Syft leave information in the SBOM to identify
that a specific component was created by them. This patch exposes the
necessary parts of the SBOM in the e2e tests mappings so that this
information can be accessed.

Syft does not add any identifying information to SPDX components, which
makes identifiying it as a creator harder. For this reason, this patch
solely focus on Hermeto components.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>