build_templates: Check for Hermeto output in the SBOM
Our e2e scenarios that cover hermetic builds make use of Hermeto to
prefetch the dependencies. This means that the produced SBOM should
contain provenance information in each of the components that were
generated by Hermeto.

This test makes sure the information is present in at least one of the
SBOM components, aiming to ensure that the SBOM generated by Hermeto is
being used in the final image SBOM composition.

STONEBLD-3872

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>