fix: Change Gitlab status report (#1716)
feat: GetCommitStatusConclusion() function

Introduce a function for getting the conclusion of checks associated
with a commit.



Get rid of ValidateNoteInMergeRequestComment()

ValidateNoteInMergeRequestComment() is used for getting the status of a
pipeline associated with a merge request.

However, pipelines-as-code no longer writes comments for pipelines that
succeeded since version 0.37.1.

Therefore, whether a pipeline has finished successfully needs to be
checked by GetCommitStatusConclusion() instead.

ValidateNoteInMergeRequestComment() is no longer necessary, remove it.

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>