chore: change default branch for integration e2e tests
The integration e2e tests trigger builds by creating a Component in the
cluster, which sends an onboarding PR to the repo/branch for that
Component in Github or Gitlab. Now that the main branches for the
Compopnents have been onboarded to the staging cluster (in order to
build the images nightly) we need to use a different branch for our main
e2e tests.  This is only necessary while there is an overlap between
migrating from the old way of running e2e tests (components are built
for every test) to the new way (compopnents are built nightly and
pulled to run integration tests).

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>